requirements.txt
requests>=2.25.1
python-docx>=0.8.11